first version diffusion API

+++ maua/super/image/models/realesrgan.py
@@ -45,5 +45,5 @@ def upscale(images: List[Union[Tensor, Image.Image, Path, str]], model):
     for img in images:
         input = load_image(img).squeeze().permute(1, 2, 0).mul(255).numpy()
         large = model.enhance(input)[0]
-        large = torch.from_numpy(large).permute(2, 0, 1).unsqueeze(0).div(255)
+        large = torch.from_numpy(large).permute(2, 0, 1).unsqueeze(0).float().div(255)
         yield large
